refactor: 결제 기반 이벤트 구조 세분화
